Slight optimization: use multiplication rather than of division.
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@6520 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKTextViewCompletionController.m
@@ -345,11 +345,11 @@ static int BDSKCompletionMinHeight = 20;
     // get the remaining space on the screen
     float hSize = NSMaxX(screenFrame) - BDSKCompletionMaxWidth - topLeftPoint.x;
     hSize = hSize <= 0.0f ? BDSKCompletionMaxWidth + hSize : BDSKCompletionMaxWidth;
-    hSize = floor(fmaxf(hSize, BDSKCompletionMinWidth));
+    hSize = floorf(fmaxf(hSize, BDSKCompletionMinWidth));
     
     float vSize = topLeftPoint.y - BDSKCompletionMaxHeight;
     vSize = vSize <= 0.0f ? BDSKCompletionMaxHeight + vSize : BDSKCompletionMaxHeight;
-    vSize = floor(fmaxf(vSize, BDSKCompletionMinHeight));
+    vSize = floorf(fmaxf(vSize, BDSKCompletionMinHeight));
     
     NSSize adjustedSize = [self windowContentSize];
     if(adjustedSize.width > hSize)
